try actually having gdb use gnulib (i.e. build-gnulib)
unfortunately the Makefile for it seems unable to understand
where $(srcdir) is, when $(srcdir) != $(builddir)

+++ src/gdb/symmisc.c
@@ -185,8 +185,9 @@ print_symbol_bcache_statistics (void)
   immediate_quit--;
 }
 
+/* */
 void
-print_objfile_statistics (void)
+print_objfile_statistics(void)
 {
   struct objfile *objfile;
   struct symtab *s;
@@ -194,33 +195,33 @@ print_objfile_statistics (void)
   int i, linetables, blockvectors;
 
   immediate_quit++;
-  ALL_OBJFILES (objfile)
+  ALL_OBJFILES(objfile)
   {
-    printf_filtered (_("Statistics for '%s':\n"), objfile->name);
-    if (OBJSTAT (objfile, n_stabs) > 0)
-      printf_filtered (_("  Number of \"stab\" symbols read: %d\n"),
-		       OBJSTAT (objfile, n_stabs));
-    if (OBJSTAT (objfile, n_minsyms) > 0)
-      printf_filtered (_("  Number of \"minimal\" symbols read: %d\n"),
-		       OBJSTAT (objfile, n_minsyms));
-    if (OBJSTAT (objfile, n_psyms) > 0)
-      printf_filtered (_("  Number of \"partial\" symbols read: %d\n"),
-		       OBJSTAT (objfile, n_psyms));
-    if (OBJSTAT (objfile, n_syms) > 0)
-      printf_filtered (_("  Number of \"full\" symbols read: %d\n"),
-		       OBJSTAT (objfile, n_syms));
-    if (OBJSTAT (objfile, n_types) > 0)
-      printf_filtered (_("  Number of \"types\" defined: %d\n"),
-		       OBJSTAT (objfile, n_types));
+    printf_filtered(_("Statistics for '%s':\n"), objfile->name);
+    if (OBJSTAT(objfile, n_stabs) > 0)
+      printf_filtered(_("  Number of \"stab\" symbols read: %d\n"),
+		      OBJSTAT(objfile, n_stabs));
+    if (OBJSTAT(objfile, n_minsyms) > 0)
+      printf_filtered(_("  Number of \"minimal\" symbols read: %d\n"),
+		      OBJSTAT(objfile, n_minsyms));
+    if (OBJSTAT(objfile, n_psyms) > 0)
+      printf_filtered(_("  Number of \"partial\" symbols read: %d\n"),
+		      OBJSTAT(objfile, n_psyms));
+    if (OBJSTAT(objfile, n_syms) > 0)
+      printf_filtered(_("  Number of \"full\" symbols read: %d\n"),
+		      OBJSTAT(objfile, n_syms));
+    if (OBJSTAT(objfile, n_types) > 0)
+      printf_filtered(_("  Number of \"types\" defined: %d\n"),
+		      OBJSTAT(objfile, n_types));
     i = 0;
-    ALL_OBJFILE_PSYMTABS (objfile, ps)
+    ALL_OBJFILE_PSYMTABS(objfile, ps)
       {
         if (ps->readin == 0)
           i++;
       }
-    printf_filtered (_("  Number of psym tables (not yet expanded): %d\n"), i);
+    printf_filtered(_("  Number of psym tables (not yet expanded): %d\n"), i);
     i = linetables = blockvectors = 0;
-    ALL_OBJFILE_SYMTABS (objfile, s)
+    ALL_OBJFILE_SYMTABS(objfile, s)
       {
         i++;
         if (s->linetable != NULL)
@@ -228,21 +229,21 @@ print_objfile_statistics (void)
         if (s->primary == 1)
           blockvectors++;
       }
-    printf_filtered (_("  Number of symbol tables: %d\n"), i);
-    printf_filtered (_("  Number of symbol tables with line tables: %d\n"),
-                     linetables);
-    printf_filtered (_("  Number of symbol tables with blockvectors: %d\n"),
-                     blockvectors);
+    printf_filtered(_("  Number of symbol tables: %d\n"), i);
+    printf_filtered(_("  Number of symbol tables with line tables: %d\n"),
+                    linetables);
+    printf_filtered(_("  Number of symbol tables with blockvectors: %d\n"),
+                    blockvectors);
 
-    if (OBJSTAT (objfile, sz_strtab) > 0)
-      printf_filtered (_("  Space used by a.out string tables: %d\n"),
-		       OBJSTAT (objfile, sz_strtab));
-    printf_filtered (_("  Total memory used for objfile obstack: %d\n"),
-		     obstack_memory_used (&objfile->objfile_obstack));
-    printf_filtered (_("  Total memory used for psymbol cache: %d\n"),
-		     bcache_memory_used (objfile->psymbol_cache));
-    printf_filtered (_("  Total memory used for macro cache: %d\n"),
-		     bcache_memory_used (objfile->macro_cache));
+    if (OBJSTAT(objfile, sz_strtab) > 0)
+      printf_filtered(_("  Space used by a.out string tables: %d\n"),
+		      OBJSTAT(objfile, sz_strtab));
+    printf_filtered(_("  Total memory used for objfile obstack: %d\n"),
+		    (int)obstack_memory_used(&objfile->objfile_obstack));
+    printf_filtered(_("  Total memory used for psymbol cache: %d\n"),
+		    bcache_memory_used(objfile->psymbol_cache));
+    printf_filtered(_("  Total memory used for macro cache: %d\n"),
+		    bcache_memory_used(objfile->macro_cache));
   }
   immediate_quit--;
 }
